Split js/shelter.js into respective files (js/aes.js and js/FileSaver.js)
